move SmtpClientExtensions to new project
